requirements.txt
openai==1.25.0
fastapi==0.105.0
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pandas==2.1.3
numpy==1.26.2
yfinance==0.2.33
ta==0.10.2
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
aiohttp==3.9.1
requests==2.31.0
httpx==0.25.2
python-jose==3.3.0
passlib==1.7.4
prometheus-client==0.19.0
python-json-logger==2.0.7
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
slowapi==0.1.8
nltk==3.8.1
beautifulsoup4==4.12.2
python-dateutil==2.8.2
pytz==2023.3.post1
tenacity==8.2.3
ujson==5.8.0
asyncio==3.4.3
backoff==2.2.1
black==23.11.0
flake8==6.1.0